requirements.txt
requests==2.32.5
feedparser==6.0.10
python-dotenv==1.0.0
schedule==1.2.0
openai==1.3.0
